fix and test save end point of product

+++ backend-luxehome/src/main/java/org/backendluxehome/modules/commun/BaseEntity.java
@@ -35,9 +35,9 @@ public class BaseEntity {
 
     @CreatedBy
     @Column(nullable = false, updatable = false)
-    private Integer createdBy;
+    private Long createdBy;
 
     @LastModifiedBy
     @Column(insertable = false)
-    private Integer lastModifiedBy;
+    private Long lastModifiedBy;
 }
